+++ R/clean.R
@@ -32,6 +32,7 @@
 #' @param trim logical to indicate whether the result should be trimmed with \code{\link{trimws}}
 #' @param ignore.case logical to indicate whether matching should be case-insensitive
 #' @param format a date format that will be passed on to \code{\link{format_datetime}}, see Details
+#' @param currency_symbol the currency symbol to use, which defaults to the current system locale setting
 #' @param ... other parameters passed on to \code{\link{as.Date}}
 #' @details
 #' Using \code{clean()} on a vector will guess a cleaning function based on the potential number of \code{NAs} it returns. Using \code{clean()} on a data.frame to apply this guessed cleaning over all columns.
@@ -42,6 +43,7 @@
 #'   \item{\code{clean_logical()}:\cr}{Use parameters \code{true} and \code{false} to match values using case-insensitive regular expressions (\link[base]{regex}). Unmatched values are considered \code{NA}. At default, values are matched with \code{\link{regex_true}} and \code{\link{regex_false}}. This allows support for values "Yes" and "No" in the following languages: Arabic, Bengali, Chinese (Mandarin), Dutch, English, French, German, Hindi, Indonesian, Japanese, Malay, Portuguese, Russian, Spanish, Telugu, Turkish and Urdu. Use parameter \code{na} to override values as \code{NA} that would else be matched with \code{true} or \code{false}. See Examples.}
 #'   \item{\code{clean_factor()}:\cr}{Use parameter \code{levels} to set new factor levels. They can be case-insensitive regular expressions to match existing values of \code{x}. For matching, new values for \code{levels} are internally temporary sorted descending on text length. See Examples.}
 #'   \item{\code{clean_numeric()} and \code{clean_character()}:\cr}{Use parameter \code{remove} to match values that must be removed from the input, using regular expressions (\link[base]{regex}). In case of \code{clean_numeric()}, comma's will be read as dots and only the last dot will be kept. Function \code{clean_character()} will keep middle spaces at default. See Examples.}
+#'   \item{\code{clean_currency()}:\cr}{This new class works like \code{clean_numeric()}, but transforms it with \code{\link{as.currency}}. See Examples.}
 #'   \item{\code{clean_Date()}:\cr}{Use parameter \code{format} to define a date format, or leave it empty to have the format guessed. Use \code{"Excel"} to read values as Microsoft Excel dates. The \code{format} parameter will be evaluated with \code{\link{format_datetime}}, which means that a format like \code{"d-mmm-yy"} with be translated internally to \code{"\%e-\%b-\%y"} for convenience. See Examples.}
 #' }
 #' 
@@ -82,6 +84,9 @@
 #' 
 #' clean_character("qwerty123456")
 #' clean_character("Positive (0.143)")
+#' 
+#' clean_currency(c("Received 25", "Received 31.40"))
+#' sum(clean_currency(c("Received 25", "Received 31.40")))
 #'  
 #' clean("12 06 2012")
 #' clean(data.frame(dates = "2013-04-02", 
@@ -201,6 +206,12 @@ clean_character <- function(x, remove = "[^a-z \t\r\n]", fixed = FALSE, ignore.c
   } else {
     x
   }
+}
+
+#' @rdname clean
+#' @export
+clean_currency <- function(x, currency_symbol = Sys.localeconv()["currency_symbol"], ...) {
+  as.currency(clean_numeric(x, ...), currency_symbol = currency_symbol)
 }